+++ lib/jekyll-crds.rb
@@ -25,4 +25,4 @@ require "jekyll-crds/filters/related_filter"
 require "jekyll-crds/filters/singularize_filter"
 require "jekyll-crds/filters/titleize_filter"
 require "jekyll-crds/filters/youtube_filter"
-require "jekyll-crds/filters/video_options_filter"
+require "jekyll-crds/filters/media_url_filter"
