+++ lib/jekyll-crds/filters/media_url_filter.rb
